[AR] Added portaudio support instead of fmod

+++ Meteora/src/modules/generators/WhiteNoise.cpp
@@ -25,7 +25,7 @@ namespace Meteora
 		// construct a trivial random generator engine from a time-based seed:
 		unsigned seed = std::chrono::system_clock::now().time_since_epoch().count();
 		std::default_random_engine generator(seed);
-		distribution = std::normal_distribution<float>(0.0, 1.0);
+		distribution = std::normal_distribution<double>(0.0, 1.0);
 	}
 
 } // namespace Meteora
